Verilator

+++ hardware/simulation/cocotb/system_tb.py
@@ -58,6 +58,8 @@ async def files_tb_test(dut):
     char = 0
     number_of_bytes_from_cnsl = 0
     number_of_bytes_from_soc = 0
+    RXready = 0
+    TXready = 0
     reset_n = dut.reset
     clk_n = dut.clk
 
@@ -69,45 +71,39 @@ async def files_tb_test(dut):
     dut.uart_wstrb.value = 0
     await inituart(dut)
 
+    soc2cnsl = open('./soc2cnsl', 'wb+')
     print('\n\nTESTBENCH: connecting')
 
     while(1):
         if(dut.trap.value.integer > 0):
             print('\nTESTBENCH: force cpu trap exit')
             break
-        RXready = 0
-        TXready = 0
         while(RXready != 1 and TXready != 1):
             RXready = await uartread(dut, UART_RXREADY_ADDR)
             TXready = await uartread(dut, UART_TXREADY_ADDR)
         if(RXready):
-            soc2cnsl = open('./soc2cnsl', 'wb+')
             if(soc2cnsl.read(1)==b''):
                 char = await uartread(dut, UART_RXDATA_ADDR)
                 soc2cnsl.write(char.to_bytes(1,  byteorder='little'))
-            soc2cnsl.close()
+                RXready = 0
+            soc2cnsl.seek(0) # absolute file positioning
         if(TXready):
             try:
                 ### IO operation ###
                 cnsl2soc = open('./cnsl2soc', 'rb+')
             except IOError as e:
                 print('Could not open file cnsl2soc!')
+                soc2cnsl.close()
                 break
             aux = cnsl2soc.read(1)
-            while(aux!=b''):
+            if(aux!=b''):
+                #print(aux, end = '')
+                #print('ENTER!')
                 send = int.from_bytes(aux, "little")
-                #print(chr(send), end = '')
-                number_of_bytes_from_cnsl += 1
-                if(number_of_bytes_from_cnsl%1000 == 0):
-                    print('-', end = '')
-                    sys.stdout.flush()
                 await uartwrite(dut, UART_TXDATA_ADDR, send)
-                TXready = 0
-                while(not TXready):
-                    TXready = await uartread(dut, UART_TXREADY_ADDR)
-                aux = cnsl2soc.read(1)
-            cnsl2soc.seek(0) # absolute file positioning
-            cnsl2soc.truncate() # to erase all data
+                cnsl2soc.seek(0) # absolute file positioning
+                cnsl2soc.truncate() # to erase all data
             cnsl2soc.close()
+            TXready = 0
 
     print('TESTBENCH: finished\n\n')
